+++ awcms/check_extension.sql
@@ -1,0 +1,2 @@
+
+SELECT * FROM extensions WHERE slug = 'mailketing';